+++ dbdaora/boolean/_tests/test_integration_service_boolean_aioredis_get_many.py
@@ -46,22 +46,20 @@ async def test_should_get_many_without_cache(
 async def test_should_get_many_from_cache(
     fake_service, serialized_fake_entity, fake_entity, fake_entity2
 ):
-    fake_service.repository.memory_data_source.exists = (
-        asynctest.CoroutineMock()
-    )
+    fake_service.repository.memory_data_source.get = asynctest.CoroutineMock()
     fake_service.cache['fake'] = fake_entity.id
     fake_service.cache['fake2'] = fake_entity2.id
     entities = await fake_service.get_many('fake', 'fake2')
 
     assert entities == [fake_entity.id, fake_entity2.id]
-    assert not fake_service.repository.memory_data_source.exists.called
+    assert not fake_service.repository.memory_data_source.get.called
 
 
 @pytest.mark.asyncio
 async def test_should_get_many_from_fallback_after_open_circuit_breaker(
     fake_service, fake_entity, fake_entity2, mocker
 ):
-    fake_service.repository.memory_data_source.exists = asynctest.CoroutineMock(
+    fake_service.repository.memory_data_source.get = asynctest.CoroutineMock(
         side_effect=RedisError
     )
     fake_service.repository.fallback_data_source.db['fake:fake'] = {
@@ -82,7 +80,7 @@ async def test_should_get_many_from_fallback_after_open_circuit_breaker_without_
     fake_service, fake_entity, fake_entity2, mocker
 ):
     fake_service.cache = None
-    fake_service.repository.memory_data_source.exists = asynctest.CoroutineMock(
+    fake_service.repository.memory_data_source.get = asynctest.CoroutineMock(
         side_effect=RedisError
     )
     fake_service.repository.fallback_data_source.db['fake:fake'] = {