+++ dbdaora/boolean/_tests/test_integration_service_boolean_aioredis_get_one.py
@@ -32,21 +32,19 @@ async def test_should_get_one_without_cache(
 async def test_should_get_one_from_cache(
     fake_service, serialized_fake_entity, fake_entity
 ):
-    fake_service.repository.memory_data_source.exists = (
-        asynctest.CoroutineMock()
-    )
+    fake_service.repository.memory_data_source.get = asynctest.CoroutineMock()
     fake_service.cache['fake'] = fake_entity.id
     entity = await fake_service.get_one('fake')
 
     assert entity == fake_entity.id
-    assert not fake_service.repository.memory_data_source.exists.called
+    assert not fake_service.repository.memory_data_source.get.called
 
 
 @pytest.mark.asyncio
 async def test_should_get_one_from_fallback_after_open_circuit_breaker(
     fake_service, fake_entity, mocker
 ):
-    fake_service.repository.memory_data_source.exists = asynctest.CoroutineMock(
+    fake_service.repository.memory_data_source.get = asynctest.CoroutineMock(
         side_effect=RedisError
     )
     fake_service.repository.fallback_data_source.db['fake:fake'] = {
@@ -64,7 +62,7 @@ async def test_should_get_one_from_fallback_after_open_circuit_breaker_without_c
     fake_service, fake_entity, mocker
 ):
     fake_service.cache = None
-    fake_service.repository.memory_data_source.exists = asynctest.CoroutineMock(
+    fake_service.repository.memory_data_source.get = asynctest.CoroutineMock(
         side_effect=RedisError
     )
     fake_service.repository.fallback_data_source.db['fake:fake'] = {