+++ dbdaora/boolean/repositories/_tests/test_integration_repository_aioredis_boolean_get.py
@@ -34,7 +34,7 @@ async def test_should_raise_not_found_error_when_already_raised_before(
 ):
     fake_entity = 'fake'
     expected_query = Query(repository, memory=True, id=fake_entity)
-    repository.memory_data_source.exists = asynctest.CoroutineMock(
+    repository.memory_data_source.get = asynctest.CoroutineMock(
         side_effect=[False, True]
     )
     repository.memory_data_source.set = asynctest.CoroutineMock()
@@ -43,9 +43,8 @@ async def test_should_raise_not_found_error_when_already_raised_before(
         await repository.query(fake_entity).entity
 
     assert exc_info.value.args == (expected_query,)
-    assert repository.memory_data_source.exists.call_args_list == [
+    assert repository.memory_data_source.get.call_args_list == [
         mocker.call('fake:fake'),
-        mocker.call('fake:not-found:fake'),
     ]
     assert not repository.memory_data_source.set.called
 
@@ -54,10 +53,10 @@ async def test_should_raise_not_found_error_when_already_raised_before(
 async def test_should_set_already_not_found_error(repository, mocker):
     fake_entity = 'fake'
     expected_query = Query(repository, memory=True, id=fake_entity)
-    repository.memory_data_source.exists = asynctest.CoroutineMock(
-        side_effect=[False, False]
-    )
     repository.fallback_data_source.get = asynctest.CoroutineMock(
+        return_value=None
+    )
+    repository.memory_data_source.get = asynctest.CoroutineMock(
         return_value=None
     )
     repository.memory_data_source.set = asynctest.CoroutineMock()
@@ -66,41 +65,36 @@ async def test_should_set_already_not_found_error(repository, mocker):
         await repository.query(fake_entity).entity
 
     assert exc_info.value.args == (expected_query,)
-    assert repository.memory_data_source.exists.call_args_list == [
-        mocker.call('fake:fake'),
-        mocker.call('fake:not-found:fake'),
-    ]
     assert repository.fallback_data_source.get.call_args_list == [
         mocker.call('fake:fake')
     ]
     assert repository.memory_data_source.set.call_args_list == [
-        mocker.call('fake:not-found:fake', '1')
+        mocker.call('fake:fake', '0')
     ]
 
 
 @pytest.mark.asyncio
 async def test_should_get_from_fallback(repository, fake_entity):
     await repository.memory_data_source.delete('fake:fake')
-    await repository.memory_data_source.delete('fake:not-found:fake')
     repository.fallback_data_source.db['fake:fake'] = {'value': True}
     entity = await repository.query(fake_entity.id).entity
 
     assert entity == fake_entity.id
-    assert repository.memory_data_source.exists('fake:fake')
+    assert await repository.memory_data_source.get('fake:fake') == b'1'
 
 
 @pytest.mark.asyncio
 async def test_should_set_memory_after_got_fallback(
     repository, fake_entity, mocker
 ):
-    repository.memory_data_source.exists = asynctest.CoroutineMock(
-        side_effect=[False, False]
+    repository.memory_data_source.get = asynctest.CoroutineMock(
+        side_effect=[None]
     )
     repository.memory_data_source.set = asynctest.CoroutineMock()
     repository.fallback_data_source.db['fake:fake'] = {'value': True}
     entity = await repository.query(fake_entity.id).entity
 
-    assert repository.memory_data_source.exists.called
+    assert repository.memory_data_source.get.called
     assert repository.memory_data_source.set.call_args_list == [
         mocker.call('fake:fake', '1')
     ]