+++ dbdaora/boolean/repositories/_tests/test_integration_repository_aioredis_boolean_get_many.py
@@ -13,8 +13,8 @@ async def test_should_set_already_not_found_error_when_get_many(
     expected_query = QueryMany(
         repository, fake_entity, memory=True, many=[fake_entity]
     )
-    repository.memory_data_source.exists = asynctest.CoroutineMock(
-        side_effect=[False, False]
+    repository.memory_data_source.get = asynctest.CoroutineMock(
+        side_effect=[None]
     )
     repository.fallback_data_source.get = asynctest.CoroutineMock(
         return_value=None
@@ -25,28 +25,26 @@ async def test_should_set_already_not_found_error_when_get_many(
         await repository.query(many=[fake_entity]).entities
 
     assert exc_info.value.args == (expected_query,)
-    assert repository.memory_data_source.exists.call_args_list == [
+    assert repository.memory_data_source.get.call_args_list == [
         mocker.call('fake:fake'),
-        mocker.call('fake:not-found:fake'),
     ]
     assert repository.fallback_data_source.get.call_args_list == [
         mocker.call('fake:fake')
     ]
     assert repository.memory_data_source.set.call_args_list == [
-        mocker.call('fake:not-found:fake', '1')
+        mocker.call('fake:fake', '0')
     ]
 
 
 @pytest.mark.asyncio
 async def test_should_get_many_from_fallback(repository, fake_entity):
     await repository.memory_data_source.delete('fake:fake')
-    await repository.memory_data_source.delete('fake:not-found:fake')
     repository.fallback_data_source.db['fake:fake'] = {'value': True}
 
     entities = await repository.query(many=[fake_entity.id]).entities
 
     assert entities == [fake_entity.id]
-    assert repository.memory_data_source.exists('fake:fake')
+    assert await repository.memory_data_source.get('fake:fake') == b'1'
 
 
 @pytest.mark.asyncio
@@ -55,14 +53,13 @@ async def test_should_get_many_with_one_item_already_not_found_from_fallback(
 ):
     await repository.memory_data_source.delete('fake:fake')
     await repository.memory_data_source.delete('fake:fake2')
-    await repository.memory_data_source.delete('fake:not-found:fake')
-    await repository.memory_data_source.set('fake:not-found:fake2', '1')
     repository.fallback_data_source.db['fake:fake'] = {'value': True}
 
     entities = await repository.query(many=[fake_entity.id, 'fake2']).entities
 
     assert entities == [fake_entity.id]
-    assert repository.memory_data_source.exists('fake:fake')
+    assert await repository.memory_data_source.get('fake:fake') == b'1'
+    assert await repository.memory_data_source.get('fake:fake2') == b'0'
 
 
 @pytest.mark.asyncio
@@ -70,11 +67,8 @@ async def test_should_get_many_with_one_item_already_not_found_and_another_not_f
     repository, fake_entity
 ):
     await repository.memory_data_source.delete('fake:fake')
-    await repository.memory_data_source.delete('fake:fake2')
+    await repository.memory_data_source.set('fake:fake2', '0')
     await repository.memory_data_source.delete('fake:fake3')
-    await repository.memory_data_source.delete('fake:not-found:fake')
-    await repository.memory_data_source.delete('fake:not-found:fake3')
-    await repository.memory_data_source.set('fake:not-found:fake2', '1')
     repository.fallback_data_source.db['fake:fake'] = {'value': True}
 
     entities = await repository.query(
@@ -82,5 +76,6 @@ async def test_should_get_many_with_one_item_already_not_found_and_another_not_f
     ).entities
 
     assert entities == [fake_entity.id]
-    assert repository.memory_data_source.exists('fake:fake')
-    assert repository.memory_data_source.exists('fake:not-found:fake3')
+    assert await repository.memory_data_source.get('fake:fake') == b'1'
+    assert await repository.memory_data_source.get('fake:fake2') == b'0'
+    assert await repository.memory_data_source.get('fake:fake3') == b'0'